Fallback to dns.lookup separately for IPv4 & IPv6
This fixes an issue where no results would be incorrectly returned
because using family:0 (the default) with dns.lookup would return only
results from /etc/hosts if any results were found there.

That meant if you had only an IPv4 match in /etc/hosts, but the OS was
capable of resolving IPv6 as well, family:0 would only return IPv4
even though family:6 could return an IPv6 address successfully.

This is fixed by querying for both separately in parallel, and combining
them to get the result that we were expecting the default family:0
option to give us in the first place.